% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{add.comment}
\alias{add.comment}
\title{Associate comment to signal}
\usage{
add.comment(comment_id, signal_id, promt = TRUE)
}
\arguments{
\item{comment_id}{numeric with comment id}

\item{signal_id}{numeric with signal id}

\item{promt}{logical \code{TRUE} to promt confirmation, \code{FALSE} to avoid confirmation prompt}
}
\description{
Add comment.
}
\examples{
\dontrun{
add.comment()
}
}
\author{
adisch
}
