% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{add.new.comment}
\alias{add.new.comment}
\title{Add new comment to comment table.}
\usage{
add.new.comment(comment, author, promt = TRUE)
}
\arguments{
\item{comment}{string with comment text}

\item{author}{string with name of the author}

\item{promt}{logical \code{TRUE} to promt confirmation, \code{FALSE} to avoid confirmation prompt}
}
\description{
Add new comment.
}
\examples{
\dontrun{
add.new.comment("offset correction to 1450mm", "adisch")
}
}
\author{
adisch
}
