% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{add.quality.flags}
\alias{add.quality.flags}
\title{Add signal quality to signal.}
\usage{
add.quality.flags(source, parameter, method, flags, from, to, author,
  promt = TRUE)
}
\arguments{
\item{source}{string with name of the source}

\item{parameter}{string with name of the parameter}

\item{method}{string with name of method}

\item{flags}{string vector with flagging}

\item{from}{date, ideally in the format \code{1900-01-01 00:00:00}}

\item{to}{date, ideally in the format \code{2017-01-01 00:00:00}}

\item{author}{string with responsible person}

\item{promt}{logical \code{TRUE} to promt confirmation, \code{FALSE} to avoid confirmation prompt}
}
\description{
Add signal quality to signal.
}
\examples{
\dontrun{
source.name = "bf_f04_23_bahnhofstr"
parameter.name = "flow rate"
method = "rangeCheck"
flags = c("green", "green", "orange", "green", "red", "green")
from = "2018-02-01 12:00:00"
to = "2018-02-01 12:25:00"
author = "adisch"
add.quality.flags(source.name, parameter.name, method, flags, tmin, tmax, author, promt = TRUE)
}
}
\author{
adisch
}
