% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{add.quality.method}
\alias{add.quality.method}
\title{Add quality for new method.}
\usage{
add.quality.method(method, flags, promt = TRUE)
}
\arguments{
\item{method}{string with name of applied method}

\item{flags}{vector of strings with flags (\code{green} madatory)}

\item{promt}{logical \code{TRUE} to promt confirmation, \code{FALSE} to avoid confirmation prompt}
}
\description{
Add quality for new method.
}
\examples{
\dontrun{
method = "rangeCheck"
flags = c("green", "orange", "red")
add.quality.method(method, flags)
}
}
\author{
adisch
}
