% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\docType{package}
\name{DatapoolR}
\alias{DatapoolR}
\alias{DatapoolR-package}
\title{DatapoolR - access the datapool of the urban water obervatory}
\description{
This little package simplifies accessing data collected at the urban water
observatory. It also provides functions to add comments.
}
\author{
Andreas Scheidegger
}
