% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{get.signal.ID}
\alias{get.signal.ID}
\title{Get a signal ID.}
\usage{
get.signal.ID(source.name, parameter.name, from = "1900-01-01 00:00:00",
  to = as.character(with_tz(now(), "UTC")))
}
\arguments{
\item{source.name}{String with name of the source}

\item{parameter.name}{String with name of the parameter}

\item{from}{date, ideally in the format \code{1900-01-01 00:00:00}}

\item{to}{date, ideally in the format \code{2017-01-01 00:00:00}}
}
\value{
data.frame with \code{time} as POSIXct and ID as \code{integer}
}
\description{
Get a signal ID.
}
\author{
adisch
}
