% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{get.source.type}
\alias{get.source.type}
\title{Get source type.}
\usage{
get.source.type(source.name)
}
\arguments{
\item{source.name}{String with name of source}
}
\value{
Integer vector with name of source type
}
\description{
Get source type of source.
}
\examples{
\dontrun{
source.name = "bf_f04_23_bahnhofstr"
get.source.type(source.name)
}
}
\author{
adisch
}
