% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{get.special.values}
\alias{get.special.values}
\title{Get special values.}
\usage{
get.special.values(source.type)
}
\arguments{
\item{source.type}{String with name of source_type}
}
\value{
Integer vector with special values
}
\description{
Get special values.
}
\examples{
\dontrun{
source.type = "Hach_Flo-Dar"
get.special.values(source.type)
}
}
\author{
adisch
}
