% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{init.db.connection}
\alias{init.db.connection}
\title{Connect to Datapool}
\usage{
init.db.connection(host = NULL, port = NULL, database = NULL,
  user = NULL, password = NULL)
}
\arguments{
\item{host}{String, specifying the database server's address}

\item{port}{Integer, specifying the port on which the database runs}

\item{database}{String, stating the database name}

\item{user}{String, specifying the database user}

\item{password}{String, specifying the database user's password}
}
\description{
This function established a connection to the datapool.
}
\details{
You must run this function after importing the DatapoolR library to connect
to the datapool database.
}
\examples{
\dontrun{
# if defaults have been set with the \\code{link[DatapoolR]{setDefaults}} function.
init.db.connection()

# connecting without default

host <- "<host-address>"
port <- <host-port>
database <- "<database-name>"
user <- "<database-user-name>"
password <- "<database-user-password>"

setDefaults(host,port,database,user,password)

}
}
\author{
Andreas Scheidegger
}
