% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{last.signals}
\alias{last.signals}
\title{Get the latest signal from each source}
\usage{
\method{last}{signals}(recent.on.top = TRUE)
}
\arguments{
\item{recent.on.top}{if \code{TRUE} the most recent signal is on
top of data.frame, if \code{FALSE} at the end.}
}
\value{
data.frame containing the signals
}
\description{
Get the latest signal from each source.
}
\examples{
\dontrun{
last.signals()
}
}
\author{
Lukas Keller
}
