% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{list.quality.flags}
\alias{list.quality.flags}
\title{List all quality methods and flags}
\usage{
list.quality.flags()
}
\value{
data.frame of all methods and flags
}
\description{
List all quality methods with the respective flags in the datapool.
}
\author{
Andy Disch
}
