% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{list.quality.methods}
\alias{list.quality.methods}
\title{List all quality methods uniquely}
\usage{
list.quality.methods()
}
\value{
data.frame of all methods
}
\description{
List all unique quality methods in the datapool.
}
\author{
Andy Disch
}
