% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{list.sites}
\alias{list.sites}
\title{List all sites}
\usage{
list.sites()
}
\value{
data.frame of all sites
}
\description{
List all sites in the datapool.
}
\author{
Andreas Scheidegger
##' \dontrun{
list.sites()
}
}
