% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{list.sources}
\alias{list.sources}
\title{List all source}
\usage{
list.sources()
}
\value{
data.frame of all sources in the datapool
}
\description{
List all types of sources in the datapool
}
\examples{
\dontrun{
list.sources()
}
}
\author{
Andreas Scheidegger
}
