% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{newest.signals}
\alias{newest.signals}
\title{Get the last signals}
\usage{
newest.signals(n, recent.on.top = TRUE)
}
\arguments{
\item{n}{number of signals}

\item{recent.on.top}{if \code{TRUE} the most recent signal is on
top of data.frame, if \code{FALSE} at the end.}
}
\value{
data.frame containing the signals
}
\description{
Get the most recent signals
}
\examples{
\dontrun{
newest.signals(4)

newest.signals(4, FALSE)
}
}
\author{
Andreas Scheidegger
}
