% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{parse.dates}
\alias{parse.dates}
\title{Parse time points}
\usage{
parse.dates(from, to)
}
\arguments{
\item{from}{date string}

\item{to}{date string}
}
\value{
list of date stings properly parsed
}
\description{
Parse time points and guess time format
}
\author{
Andreas Scheidegger
}
