% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR_add.r
\name{PlotDatapoolR}
\alias{PlotDatapoolR}
\title{DatapoolR plots}
\usage{
PlotDatapoolR(signal, show.red = TRUE)
}
\arguments{
\item{signal}{data.frame containing the signals, \code{DatapoolR} structure}

\item{show.red}{logical if red values are shown \code{TRUE} (as default) or not \code{FALSE}}
}
\value{
list.of.graphs a list of dygraph objects
}
\description{
idea: generic plotting function to visualize DatapoolR outputs
}
\examples{
\dontrun{
PlotDatapoolR(signal)
}
}
\author{
adisch
}
