% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.r
\name{removeDefaults}
\alias{removeDefaults}
\title{Remove Defaults}
\usage{
removeDefaults()
}
\value{
String, stating if removing was possible.
}
\description{
This function is able to remove previously set default connection parameters.
}
\author{
Christian Foerster
}
