% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.r
\name{setDefaults}
\alias{setDefaults}
\title{Set Defaults}
\usage{
setDefaults(host, port, database, user, password)
}
\arguments{
\item{host}{String, specifying the database server's address}

\item{port}{Integer, specifying the port on which the database runs}

\item{database}{String, stating the database name}

\item{user}{String, specifying the database user}

\item{password}{String, specifying the database user's password}
}
\value{
String, stating if setting was possible.
}
\description{
This function is able to set default connection parameters.
}
\author{
Christian Foerster
}
