% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{signals.at.site}
\alias{signals.at.site}
\title{Get signals from a site}
\usage{
signals.at.site(site.name, from = "1900-01-01 00:00:00",
  to = "2019-01-01 00:00:00")
}
\arguments{
\item{site.name}{string of the site name}

\item{from}{date, ideally in the format \code{1900-01-01 00:00:00}}

\item{to}{date, ideally in the format \code{2017-01-01 00:00:00}}
}
\value{
data.frame containing the signals
}
\description{
Get all signals measured at a given site.
}
\examples{
\dontrun{
site = "mysite"
from <- "2016-06-22 12"        # zeros are automatically added
to <- "2016.06.22 12:10:00"  # parser tries to guess the datetime format
signals.at.site(site, from, to)
}
}
\author{
Andreas Scheidegger
}
