% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{signals.of.source}
\alias{signals.of.source}
\title{Get all or defined signals from a source instance}
\usage{
signals.of.source(source.name, from = "1900-01-01 00:00:00",
  to = "2019-01-01 00:00:00", parameter.name = NULL)
}
\arguments{
\item{source.name}{name of source instance}

\item{from}{date, ideally in the format \code{1900-01-01 00:00:00}}

\item{to}{date, ideally in the format \code{2017-01-01 00:00:00}}

\item{parameter.name}{string of parameter name (optional, \code{NULL} by default)}
}
\value{
data.frame containing the signals
}
\description{
Get all or defined signals from a source instance.
}
\examples{
\dontrun{
source.name = "bf_f04_23_bahnhofstr"
from = "2018-07-11 00:00:00"
to = "2018-07-11 23:55:00"
parameter.name <- "flow rate"
signals.of.source(source.name, from, to, parameter.name)
}
}
\author{
Andreas Scheidegger, Andy Disch
}
