% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapoolR.r
\name{signals.of.sourcetype}
\alias{signals.of.sourcetype}
\title{Get all signals from a source type}
\usage{
signals.of.sourcetype(source.type, from = "1900-01-01 00:00:00",
  to = "2019-01-01 00:00:00")
}
\arguments{
\item{source.type}{name of source type}

\item{from}{date, ideally in the format \code{1900-01-01 00:00:00}}

\item{to}{date, ideally in the format \code{2017-01-01 00:00:00}}
}
\value{
data.frame containing the signals
}
\description{
Get all signals from a source type.
}
\examples{
\dontrun{
source.type = "Hach_Flo-Dar"
signals.of.sourcetype(source.name)
}
}
\author{
Andreas Scheidegger
}
